230. Kth Smallest Element in a BST
